\name{geesibsor-internal}
\alias{gee.sibs.or.bits}

\title{Internal geesibsor functions}

\description{
  Internal geesibsor functions.
}

\usage{
gee.sibs.or.bits(y, x, beta, gamma, maxit=1000, eta.tol=1e-9, trace=FALSE)
}

\details{
  These are generally not to be called by the user.
}

\author{Karl W Broman, \email{kbroman@biostat.wisc.edu} }

\keyword{internal}
